法廊/lawundry/requirement.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
Babel==2.15.0
backend==0.2.4.1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cs50==9.3.4
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
docker-pycreds==0.4.0
executing==2.0.1
ExifRead==3.0.0
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-SSLify==0.1.5
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==4.41.0
gradio_client==1.3.0
grpcio==1.65.5
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
imageio==2.35.0
importlib_resources==6.4.3
inifile==0.4.1
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
Lektor==3.3.11
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==0.8.4
ml-dtypes==0.4.0
mpmath==1.3.0
msgspec==0.18.6
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
pkg==0.2
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.58
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
ruff==0.6.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.10
tomlkit==0.12.0
torch
torchvision
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.64
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.3
wandb==0.17.7
watchdog==4.0.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==2.3.8
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.2
